This is the fixed HTML file.
I was on of the students that had a Chromebook that wasn't compatible with Atom or GitKraken. The Karot application you mentioned on black board was not compatible with my old Chromebook. I just got a Mac Book today and did the assignment. I am having a hard time with installing GitKraken. I downloaded it but the link to launch the app has a question mark behind it and the application will not launch.